fix: remove unnecessary type casts
- Remove unnecessary u64 -> u64 cast in multi_peer.rs hasher.finish()
- Remove unnecessary u32 -> u32 cast in multi_peer.rs peer_height
- Remove unnecessary u32 -> u32 cast in network/mod.rs height

These types were already the correct type, so casting them was
unnecessary and created avoidable runtime overhead. Fixes 3 clippy
unnecessary_cast warnings.

+++ dash-spv/src/network/mod.rs
@@ -302,7 +302,7 @@ impl NetworkManager for TcpNetworkManager {
         if let Some(connection) = &self.connection {
             // For single peer connection, return the peer's best height
             match connection.peer_info().best_height {
-                Some(height) if height > 0 => Ok(Some(height as u32)),
+                Some(height) if height > 0 => Ok(Some(height)),
                 _ => Ok(None),
             }
         } else {
